requirements.txt
Flask
python-dotenv
gunicorn
google-genai>=0.15.0 
pymongo>=4.6
flask-cors

